split kafka and pulsar sub-generators
