+++ generators/spring-cloud-stream/generators/kafka/cleanup.ts
