+++ generators/spring-cloud-stream/generators/kafka/templates/buildSrc/src/main/groovy/jhipster.kafka-conventions.gradle.ejs
