+++ generators/spring-cloud-stream/generators/kafka/templates/src/test/java/_package_/config/EmbeddedKafka.java.ejs
