+++ generators/spring-cloud-stream/generators/pulsar/templates/buildSrc/src/main/groovy/jhipster.pulsar-conventions.gradle.ejs
