+++ generators/spring-cloud-stream/generators/pulsar/templates/src/test/java/_package_/broker/PulsarIT.java.ejs
