+++ generators/spring-cloud-stream/generators/pulsar/templates/src/test/java/_package_/config/EmbeddedPulsar.java.ejs
